Use published image rather than copying the source
This switches to using the image published by Professional Wiki, rather than copying the source.

https://github.com/ProfessionalWiki/mw-docker/blob/master/mw37/Dockerfile

This way you can still extend the image but will get updates to the base image automatically. As it stands, the copy is already out of date, since the original got updated earlier today to use the latest patch release of MediaWiki.

Of course if you wish to maintain a copy, you can do that. Though then it would be good to remove the "Professional Wiki version of base MediaWiki" source comment